sophia_turtle::serializer::trig

Struct TrigSerializer

pub struct TrigSerializer<W> { /* private fields */ }

Trig serializer.

Implementations

impl<W> TrigSerializer<W>
where W: Write,

pub fn new(write: W) -> TrigSerializer<W>

Build a new Trig serializer writing to write, with the default config.

pub fn new_with_config(write: W, config: TrigConfig) -> TrigSerializer<W>

Build a new Trig serializer writing to write, with the given config.

pub fn config(&self) -> &TrigConfig

Borrow this serializer’s configuration.

impl TrigSerializer<Vec<u8>>

pub fn new_stringifier() -> Self

Create a new serializer which targets a String.

pub fn new_stringifier_with_config(config: TrigConfig) -> Self

Create a new serializer which targets a String with a custom config.