sophia_turtle::serializer::turtle

Struct TurtleConfig

pub struct TurtleConfig { /* private fields */ }

Turtle serializer configuration.

Implementations

impl TurtleConfig

pub fn pretty(&self) -> bool

Should the parser make extra effort to produce pretty Turtle.

If false (default), the triples will be serialized in streaming mode. Subject and predicate โfactorizationโ will only occur based on the previous triple(s) in the stream. The collection syntax for rdf:Lists will not be used.

If true, extra effort will be made to group related triples together, and to use the collection syntax whenever possible. This requires storing the whole graph in memory.

pub fn prefix_map(&self) -> &[PrefixMapPair]

PrefixMap to use in serialization. (defaults to a map containing rdf:, rdfs: and xsd:)

NB: currently, only used if pretty is true.

pub fn indentation(&self) -> &str

Indentation to use in serialization. (defaults to " ", can only contain ASCII whitespaces)

NB: currently, only used if pretty is true.

pub fn new() -> Self

Build a new default TurtleConfig.

pub fn with_pretty(self, b: bool) -> Self

Transform a TurtleConfig by setting the pretty flag.

pub fn with_prefix_map<P: PrefixMap + ?Sized>(self, pm: &P) -> Self

Transform a TurtleConfig by setting the prefix_map flag (copying pm using PrefixMap::to_vec).

pub fn with_own_prefix_map(self, pm: Vec<PrefixMapPair>) -> Self

Transform a TurtleConfig by setting the prefix_map flag.

pub fn with_indentation<T: ToString>(self, indentation: T) -> Self

Transform a TurtleConfig by setting the indentation flag.

Precondition

indentation must only contain ASCII whitespaces, otherwise this method will panic.

pub fn default_prefix_map() -> Vec<PrefixMapPair>

Return the prefix map that is used when none is provided